+++ RELEASE_NOTES.md
@@ -1,3 +1,13 @@
+# BetterBlizzFrames 2.0.1c
+## Classics (Era, TBC & MoP)
+### New
+- "Larger Frames" setting for default Party Frames (Non-raid). Makes them similar to TargetFrame. On by default while "Bigger Healthbars" is enabled and optionally under PartyFrames in /bbf otherwise.
+- "Cast Target Text" setting in Castbars section. This shows name of true target of spell cast inside the castbar. Same as on retail just ported to classics now.
+- "Cast On Me Highlight" setting in Castbars section. Red highlight around castbar when the cast is being casted on you (true target, no macro workaround)
+## Midnight
+### Bugfix
+- Fix castbars showing normal cast/channel texture on an uninterruptible cast with certain specific settings.
+
 # BetterBlizzFrames 2.0.1b
 ## Retail
 ### Tweak